# src/floodcat/api/schemas.py
from typing import Any
from pydantic import BaseModel, Field, ConfigDict

class AnalysisRequest(BaseModel):
    model_config=ConfigDict(extra='forbid')
    rows: list[dict[str,Any]]=Field(min_length=1,max_length=10000)
    allow_partial: bool=False
    enhanced: bool=False
    config: dict[str,Any] | None=None

class EvidenceRequest(BaseModel):
    model_config=ConfigDict(extra='forbid')
    evidence_id: str=Field(min_length=1,max_length=100)
    source: str=Field(min_length=1,max_length=2000)
    quote: str=Field(min_length=1,max_length=10000)
    event_date: str
    location_name: str=Field(min_length=1,max_length=200)
    lat: float=Field(ge=-90,le=90,allow_inf_nan=False)
    lon: float=Field(ge=-180,le=180,allow_inf_nan=False)
    confidence: float=Field(ge=0,le=1,allow_inf_nan=False)
    drainage_signal: float=Field(ge=0,le=1,allow_inf_nan=False)

class ApprovalRequest(BaseModel):
    reviewer: str=Field(min_length=1,max_length=200)

class ExtractionRequest(BaseModel):
    text: str=Field(min_length=1,max_length=50000)
    source: str=Field(min_length=1,max_length=2000)

class CSVAnalysisRequest(BaseModel):
    model_config=ConfigDict(extra='forbid')
    csv_text: str=Field(min_length=1,max_length=10_000_000)
    allow_partial: bool=False
    enhanced: bool=False
    config: dict[str,Any] | None=None